Removed AbstractRepositoryManager
